ResultScreen Refactoring
Now it's in MVC structure
