+++ Assets/Scripts/ResultScreen/RSApplication.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+// ResultScreen Application Class for combining the MVC classes
+public class RSApplication : MonoBehaviour
+{
+    public RSModel model;
+    public RSView view;
+    public RSController controller;
+}
