+++ Assets/Scripts/ResultScreen/RSController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5e19275a30bb24a9f83a52fb3b89a8af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
