+++ Assets/Scripts/ResultScreen/RSElements.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+// Helper function for the MVC classes of ResultScreen
+public class RSElements : MonoBehaviour
+{
+   public RSApplication app {get {return GameObject.FindObjectOfType<RSApplication>();}}
+}
