+++ Assets/Scripts/ResultScreen/RSElements.cs.meta
@@ -1,9 +1,5 @@
 fileFormatVersion: 2
-<<<<<<< HEAD
-guid: 188807bb223e341e982bbd621d0875ab
-=======
-guid: c59bc815c7480473798e982659e8c793
->>>>>>> adhella0.2.2
+guid: 3eec9b8c4611d4f62ac282c34159fdc4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
