+++ Assets/Scripts/ResultScreen/RSModel.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d8a604a9d754f46f4acee2ec6c2e3e0d
+guid: ac53951b10f6142cf99b038cdb6ddabe
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
