+++ Assets/Scripts/ResultScreen/RSView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 50d1017f995d443558e00a36d6ba7952
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
